Added NewsDiffUtilCallback and TimeUtils.

+++ app/src/main/java/ru/job4j/cryptocompareapp/repository/database/entity/News.kt
@@ -20,7 +20,7 @@ data class News(
 
     @SerializedName("published_on")
     @Expose
-    val publishedOn: Int? = null,
+    val publishedOn: Long? = null,
 
     @SerializedName("imageurl")
     @Expose
